requirements.txt
argcomplete==1.10.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
blinker==1.4
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
Flask==1.1.1
Flask-API==2.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
flask-swagger-ui==3.25.0
Flask-User==1.0.2.2
Flask-WTF==0.14.2
IMAPClient==2.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.5.0
MarkupSafe==1.1.1
olefile==0.46
passlib==1.7.2
pdfminer.six==20181108
Pillow==7.1.1
pycparser==2.19
pycryptodome==3.9.7
python-pptx==0.6.18
pytz==2019.3
pywin32==227
six==1.12.0
sortedcontainers==2.1.0
soupsieve==2.0
SpeechRecognition==3.8.1
SQLAlchemy==1.3.11
textract==1.6.3
tzlocal==1.5.1
Werkzeug==0.16.0
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.8
